Bylaws - Amendment 1
In accordance with ARTICLE XIII, Section 1, a vote at the annual meeting of Association members on March 1, 2000 at Sunrise Elementary School, voting 123 to 0, recommends revision of the method of verification of the Association books by a public accountant as defined in ARTICLE VIII, Section 8. (d). from an audit to a review.